% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_module.R
\name{create_module}
\alias{create_module}
\title{Creates module of the specified name}
\usage{
create_module(name, pkg = get_wd(), open = TRUE, dir_create = TRUE)
}
\description{
Creates module of the specified name
}
