% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyHLTH_example.R
\name{shinyHLTH_example}
\alias{shinyHLTH_example}
\title{Launch an example dashboard}
\usage{
shinyHLTH_example(example = "demo_dash")
}
\description{
A gallery of widgets available in the package.
}
\examples{
if (interactive()) {

 shinyHLTH_example()

}
}
