% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard_new.R, R/start_dashboard.R
\name{start_dashboard}
\alias{start_dashboard}
\title{Golem dashboard backend}
\usage{
start_dashboard(path, ...)

start_dashboard(path, ...)
}
\description{
Creates golem directory and header from template info

Creates golem directory and header from template info
}
